Changing the name of the option.
